Some minor additions
Captain's log: today was a bad day. Couldn't really focus and could only make some minor additions.

I've been pondering the curses. On the one hand, we know they'll lead to that the lands are fucked. On the other hand: it wouldn't make sense if the gods or Emperor McBastard would curse the land directly. The gods curse him to punish him for his actions. He curses the gods to spit them. I've tried to write the curses to sort of  set up things. May have to reconsider three curses on the land? Three times three curses would be the most narratively appropriate, but maybe I should just stop trying to make that happen and stick to one really good one.

I like the idea of the curse spreading as a way for why people started coming here a lot, maybe more so than just 'here be treasure.'

We need some sort of town; a place where players can start out, and consider a safe base, maybe.  I must find some narrative reason for why there are more than just 'forward bases'. Regular joes are usefu and I'd rather not NOT have them here.
It also sort of implies though, that the adventure/colonisation of the Gloomlands has been going on for a while. Towns/villages take a while to form.